Implement template literals.
JerryScript-DCO-1.0-Signed-off-by: Zoltan Herczeg zherczeg.u-szeged@partner.samsung.com

+++ jerry-core/parser/js/js-parser-statm.c
@@ -588,7 +588,7 @@ parser_parse_with_statement_end (parser_context_t *context_p) /**< context */
 
     if (type == PARSER_STATEMENT_START)
     {
-      context_p->status_flags &= ~PARSER_INSIDE_WITH;
+      context_p->status_flags &= (uint32_t) ~PARSER_INSIDE_WITH;
       return;
     }
